backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.3.0
openai-whisper>=20231117
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.26.4
opencv-python>=4.9.0
mediapipe>=0.10.9
google-generativeai>=0.4.0
# HTTP client for LLM API
aiohttp>=3.9.0

# Video/Audio Processing
moviepy>=2.0.0
pydub>=0.25.1
numpy>=1.24.0

# Visual Analysis
opencv-python>=4.8.0
mediapipe>=0.10.0

# Utilities
pillow>=10.0.0